desafio_2/requirements.txt
Django>=4.2.1,<4.3
psycopg2-binary>=2.9.6,<2.10
pika>=1.3.2,<1.3.5
djangorestframework>=3.14.0,<3.15.0
Faker==10.0.0